SERVER-9526 alternative solution
A more elegant solution to printout of BinData in log messages. This commit changes the behaviour of BSONElement::toString() to print BinData in the same way as the other element types. It puts a cap on the lenght of the binary data in the same way as for string. Using full == true overrides the cap in the same way.